Consolidate `Schema` and `Spec` component paths
