+++ gatsby/src/components/Schema/Content/Content.module.scss
