+++ gatsby/src/components/Schema/Content/Properties.module.scss
