+++ gatsby/src/components/Schema/Heading.tsx
@@ -2,9 +2,9 @@ import { SchemaObject } from 'openapi3-ts';
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import { Type } from '../Spec';
-import { getRefAnchorLink } from '../Spec/utils';
 import styles from './Heading.module.scss';
+import Type from './Type';
+import { getRefAnchorLink } from './utils';
 
 interface IHeading {
   name: string;
