+++ gatsby/src/components/Schema/Type.module.scss
