+++ gatsby/src/components/Schema/Type.tsx
