+++ gatsby/src/components/Schema/utils.tsx
