--- gatsby/src/components/Spec/index.ts
@@ -1,2 +1,0 @@
-export { default as Properties } from './Properties';
-export { default as Type } from './Type';
